+++ packages/filter/src/lib/components/expandedFilterGroup/expandedFilterGroup.styles.ts
